restructure folders to allow database implementation
